refactor(models): create HLO subdirectory
Eventually, all models under that directory will be replaced by models
using an NxD compatible modeling.
