+++ optimum/neuron/models/inference/hlo/phi3/modules.py
@@ -14,7 +14,7 @@
 # ==============================================================================
 from transformers.models.phi import PhiConfig
 
-from ...backends.hlo import module
+from .....backends.hlo import module
 
 
 class Phi3ForCausalLM(module.PretrainedModel):
