+++ optimum/neuron/models/inference/hlo/qwen2/modules.py
@@ -16,7 +16,7 @@
 
 from transformers.models.qwen2 import Qwen2Config
 
-from ...backends.hlo import module
+from .....backends.hlo import module
 from ..llama.modules import LlamaMLP, LlamaRMSNorm
